ANDROID: Use small core data structures for microdroid kernels
CONFIG_BASE_FULL governs the size of several data structures in the
system, such as futex queue hash table size, UDP hash tables size,
and the max number of PIDs in the system. When it is enabled, the kernel
allocates larger data structures, such as a larger futex queue hash
table. When it is not enabled, the kernel uses a smaller amount of
memory for the same data structures.

Microdroid is meant to host a small number of processes, so remove
CONFIG_BASE_FULL to save memory.

Bug: 434256355
Change-Id: Id8d9957ad16377334aa321a77568e97d7e433330
Signed-off-by: Isaac J. Manjarres <isaacmanjarres@google.com>