Revert "ANDROID: gic: Add affinity-setting vendor hooks to gic-v3"
This reverts commit a2a0550225ce547d1e3d80e40d180affcc4e0cce.

The hook android_vh_gic_v3_affinity_init is not used by any vendor, so remove
it to help with merge issues with future LTS releases.

If this is needed by any real user, it can easily be reverted to add it
back and then the symbol should be added to the abi list at the same
time to prevent it from being removed again later.

Bug: 203756332
Bug: 266450121
Cc: Guru Das Srinagesh <quic_gurus@quicinc.com>
Change-Id: Id7603b35e86f09936c25f1a3fd76ef3d6d5a8308
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>